more code polish
